+++ source/helpers.js
@@ -1,9 +1,8 @@
 const fs = require('fs');
-const { Collection, TextChannel, ChannelManager, GuildChannelManager, Message, MessageOptions, Guild, ChannelType, PermissionsBitField, ActionRowBuilder, ButtonBuilder, GuildScheduledEventEntityType, StringSelectMenuBuilder, ButtonStyle, GuildMember } = require('discord.js');
+const { Collection, TextChannel, ChannelManager, GuildChannelManager, Message, MessageOptions, Guild, PermissionsBitField, ActionRowBuilder, ButtonBuilder, GuildScheduledEventEntityType, StringSelectMenuBuilder, ButtonStyle, MessagePayload, MessageFlags } = require('discord.js');
 const { Club, ClubTimeslot } = require('./classes/Club');
-const { MAX_SET_TIMEOUT } = require('./constants');
+const { MAX_SET_TIMEOUT, SAFE_DELIMITER } = require('./constants');
 const { embedTemplateBuilder, clubEmbedBuilder } = require('./engines/messageEngine');
-const { getTopicIds, getTopicNames, addTopic } = require('./engines/channelEngine');
 
 /** Convert an amount of time from a starting unit to a different one
  * @param {number} value
@@ -88,20 +87,22 @@ exports.removeModerator = function (removedId) {
 	exports.saveModData();
 }
 
-exports.modRoleId = require("../config/auth.json").modRoleId;
-
-exports.noAts = require("../config/modData.json").noAts; // [userId]
+const { noAts, modRoleId } = require("../config/modData.json");
+/** @type {string} */
+exports.modRoleId = modRoleId;
+/** @type {string[]} */
+exports.noAts = noAts;
 
 exports.atIds = new Set(); // contains userIds
 //#endregion
 
-// {[type]: {messageId: string, channelId: string}}
+/** @type {{petition: {channelId: string; messageId: string}, club: {channelId: string; messageId: string;}}} */
 exports.listMessages = require('../config/listMessageIds.json');
 
+/**  key: topic, value: petitioner ids
+ * @type {Record<string, string[]>} */
 let petitions = require('../config/petitionList.json');
-/** Get the dictionary relating topic petitions to their arrays of petitioner ids
- * @returns {Record<string, string[]>} Record<petition, petitionerId[]>
- */
+
 exports.getPetitions = function () {
 	return petitions;
 }
@@ -113,7 +114,7 @@ exports.getPetitions = function () {
 exports.setPetitions = function (petitionListInput, channelManager) {
 	petitions = petitionListInput;
 	exports.saveObject(petitions, 'petitionList.json');
-	exports.updateList(channelManager, "topics");
+	exports.updateList(channelManager, "petition");
 }
 
 const clubDictionary = {};
@@ -143,224 +144,95 @@ exports.updateClub = function (club) {
 exports.removeClub = function (id, channelManager) {
 	delete clubDictionary[id];
 	exports.saveObject(clubDictionary, 'clubList.json');
-	exports.updateList(channelManager, "clubs");
+	exports.updateList(channelManager, "club");
 }
 
-// {[textChannelId]: timeout}
+/** @type {{[clubId: string]: NodeJS.Timeout}} */
 exports.reminderTimeouts = {};
 
-// {[voiceChannelId]: timeout}
-exports.eventTimeouts = {};
-
 // Functions
-/** Update the club or topics list message
- * @param {GuildChannelManager} channelManager
- * @param {"topics" | "clubs"} listType
- * @returns {Promise<Message>}
+/** Builds the MessageOptions for the specified list message
+ * @param {ChannelManager} channelManager
+ * @param {"petition" | "club"} listType
+ * @returns {Promise<MessagePayload>}
  */
-exports.updateList = async function (channelManager, listType) {
-	const { channelId, messageId } = exports.listMessages[listType];
-	if (channelId && messageId) {
-		const channel = await channelManager.fetch(channelId);
-		const message = await channel.messages.fetch(messageId);
-		const messageOptions = await (listType == "topics" ? exports.topicListBuilder : exports.clubListBuilder)(channelManager);
-		message.edit(messageOptions);
-		if (messageOptions.files.length === 0) {
-			message.removeAttachments();
-		}
-		return message;
-	}
-}
+exports.buildListMessagePayload = function (channelManager, listType) {
+	let description;
 
-/** Create the ActionRowBuilder containing the selects for joining clubs/topics and adding to petitions
- * @param {"topics" | "petitions" | "clubs"} listType
- * @returns {ActionRowBuilder}
- */
-function listSelectBuilder(listType) {
-	let selectCutomId = "";
-	let placeholderText = "";
-	const entries = [];
+	const messageOptions = {
+		flags: MessageFlags.SuppressNotifications,
+		components: [new ActionRowBuilder()]
+	}
+
+	const selectMenu = new StringSelectMenuBuilder().setCustomId(`${listType}List`)
+		.setMinValues(1);
 
 	switch (listType) {
-		case "topics":
-			selectCutomId = "topicList";
+		case "petition":
+			description = `Here are the topic channels that have been petitioned for. Note that petitions will be converted to lowercase to match with Discord text channels being all lowercase. They will automatically be added when reaching **${Math.ceil(channelManager.guild.memberCount * 0.05)} petitions** (5% of the server). You can sign onto an already open petition with the select menu under this message (jump to message in pins).\n`;
 
-			let topicNames = getTopicNames();
-			let topicIds = getTopicIds();
-			for (let i = 0; i < topicNames.length; i++) {
-				entries.push({
-					label: topicNames[i],
-					value: topicIds[i]
-				})
+			for (const topicName in petitions) {
+				description += `\n${topicName}: ${petitions[topicName].length} petitioner(s) so far`;
+				selectMenu.addOptions([{
+					label: topicName,
+					value: topicName
+				}]);
 			}
 
-			if (entries.length > 0) {
-				placeholderText = "Select a topic...";
+			if (selectMenu.options.length > 0) {
+				selectMenu.setPlaceholder("Select a petition...");
 			} else {
-				placeholderText = "No topics yet";
+				selectMenu.setPlaceholder("No open petitions");
 			}
 			break;
-		case "petitions":
-			selectCutomId = "petitionList";
-			for (const petition in exports.getPetitions()) {
-				entries.push({
-					label: petition,
-					value: petition
-				})
-			}
+		case "club":
+			description = "Here's a list of the clubs on the server. Learn more about one by typing `/club-invite (club ID)`.\n";
 
-			if (entries.length > 0) {
-				placeholderText = "Select a petition...";
-			} else {
-				placeholderText = "No open petitions";
-			}
-			break;
-		case "clubs":
-			selectCutomId = "clubList";
-
-			for (const club of Object.values(exports.getClubDictionary())) {
+			for (const id in clubDictionary) {
+				const club = clubDictionary[id];
+				description += `\n__**${club.title}**__ (${club.userIds.length}${club.seats !== -1 ? `/${club.seats}` : ""} Members)\n**ID**: ${club.id}\n**Host**: <@${club.hostId}>\n`;
+				if (club.system) {
+					description += `**Game**: ${club.system}\n`;
+				}
+				if (club.timeslot.nextMeeting) {
+					description += `**Next Meeting**: <t:${club.timeslot.nextMeeting}>${club.timeslot.periodCount === 0 ? "" : ` repeats every ${club.timeslot.periodCount} ${club.timeslot.periodUnits === "weeks" ? "week(s)" : "day(s)"}`}\n`;
+				}
 				if (club.isRecruiting()) {
-					entries.push({
-						label: club.title,
-						description: `${club.userIds.length}${club.seats !== -1 ? `/${club.seats}` : ""} Members`,
-						value: club.id
-					});
+					selectMenu.addOptions([
+						{
+							label: club.title,
+							description: `${club.userIds.length}${club.seats !== -1 ? `/${club.seats}` : ""} Members`,
+							value: club.id
+						}
+					])
 				}
 			}
 
-			if (entries.length > 0) {
-				placeholderText = "Get club details...";
+			if (selectMenu.options.length > 0) {
+				selectMenu.setPlaceholder("Get club details...");
 			} else {
-				placeholderText = "No clubs currently recruiting";
+				selectMenu.setPlaceholder("No clubs currently recruiting");
 			}
 			break;
 	}
 
-	return new ActionRowBuilder().addComponents(
-		new StringSelectMenuBuilder()
-			.setCustomId(selectCutomId)
-			.setPlaceholder(placeholderText)
-			.setDisabled(entries.length < 1)
-			.addOptions(entries.length > 0 ? entries : [{
+	if (selectMenu.options.length > 0) {
+		selectMenu.setMaxValues(selectMenu.options.length);
+	} else {
+		selectMenu.setDisabled(true)
+			.addOptions([{
 				label: "no entries",
 				value: "no entries"
 			}])
-			.setMinValues(1)
-			.setMaxValues(entries.length > 0 ? entries.length : 1)
-	);
-}
-
-/** Create the MessageOptions for the topic list message
- * @param {GuildChannelManager} channelManager
- * @returns {Promise<MessageOptions>}
- */
-exports.topicListBuilder = function (channelManager) {
-	const messageOptions = {};
-
-	// Select Menus
-	messageOptions.components = [listSelectBuilder("topics"), listSelectBuilder("petitions")];
-
-	// Generate Message Body
-	let description = "Here's a list of the opt-in topic channels for the server. Join them by using `/join` or by using the select menu under this message (jump to message in pins).\n";
-	let topics = getTopicIds();
-
-	for (let i = 0; i < topics.length; i += 1) {
-		let id = topics[i];
-		let channel = channelManager.resolve(id);
-		if (channel) {
-			description += `\n__${channel.name}__${channel.topic ? ` ${channel.topic}` : ""}`;
-		}
+			.setMaxValues(1);
 	}
 
-	let petitionNames = Object.keys(petitions);
-	let petitionText = `Here are the topic channels that have been petitioned for. Note that petitions will be converted to lowercase to match with Discord text channels being all lowercase. They will automatically be added when reaching **${Math.ceil(channelManager.guild.memberCount * 0.05)} petitions** (5% of the server). You can sign onto an already open petition with the select menu under this message (jump to message in pins).\n`;
-	if (petitionNames.length > 0) {
-		petitionNames.forEach(topicName => {
-			petitionText += `\n${topicName}: ${petitions[topicName].length} petitioner(s) so far`;
-		})
-	}
-
-	if (description.length > 2048 || petitionText.length > 1024) {
-		return new Promise((resolve, reject) => {
-			let fileText = description;
-			if (petitionNames.length > 0) {
-				fileText += `\n\n${petitionText}`;
-			}
-
-			fs.writeFile("data/TopicChannels.txt", fileText, "utf8", error => {
-				if (error) {
-					console.error(error);
-				}
-			});
-			resolve(messageOptions);
-		}).then(() => {
-			messageOptions.embeds = [];
-			messageOptions.files = [{
-				attachment: "data/TopicChannels.txt",
-				name: "TopicChannels.txt"
-			}];
-			return messageOptions;
-		})
-	} else {
-		return new Promise((resolve, reject) => {
-			let embed = embedTemplateBuilder()
-				.setTitle("Topic Channels")
-				.setDescription(description)
-				.setFooter({ text: "Please do not make bounties to vote for your petitions." });
-
-			if (petitionNames.length > 0) {
-				embed.addFields({ name: "Petitioned Channels", value: petitionText })
-			}
-			messageOptions.embeds = [embed];
-			messageOptions.files = [];
-			resolve(messageOptions);
-		})
-	}
-}
-
-/** Pin the topics list message
- * @param {GuildChannelManager} channelManager
- * @param {TextChannel} channel
- */
-exports.pinTopicsList = function (channelManager, channel) {
-	exports.topicListBuilder(channelManager).then(messageOptions => {
-		channel.send(messageOptions).then(message => {
-			exports.listMessages.topics = {
-				"messageId": message.id,
-				"channelId": message.channelId
-			}
-			exports.saveObject(exports.listMessages, "listMessageIds.json");
-			message.pin();
-		})
-	}).catch(console.error);
-}
-
-/** Create the MessageOptions for the club list message
- * @returns {Promise<MessageOptions>}
- */
-exports.clubListBuilder = function () {
-	const messageOptions = {};
-
-	messageOptions.components = [listSelectBuilder("clubs")];
-
-	let description = "Here's a list of the clubs on the server. Learn more about one by typing `/club-invite (club ID)`.\n";
-	let clubs = exports.getClubDictionary();
-
-	Object.keys(clubs).forEach(id => {
-		let club = clubs[id];
-		description += `\n__**${club.title}**__ (${club.userIds.length}${club.seats !== -1 ? `/${club.seats}` : ""} Members)\n**ID**: ${club.id}\n**Host**: <@${club.hostId}>\n`;
-		if (club.system) {
-			description += `**Game**: ${club.system}\n`;
-		}
-		if (club.timeslot.nextMeeting) {
-			description += `**Next Meeting**: <t:${club.timeslot.nextMeeting}>${club.timeslot.periodCount === 0 ? "" : ` repeats every ${club.timeslot.periodCount} ${club.timeslot.periodUnits === "weeks" ? "week(s)" : "day(s)"}`}\n`;
-		}
-	})
+	messageOptions.components[0].addComponents(selectMenu);
 
 	if (description.length > 2048) {
 		return new Promise((resolve, reject) => {
 			let fileText = description;
-			fs.writeFile("data/ClubChannels.txt", fileText, "utf8", error => {
+			fs.writeFile("data/listMessage.txt", fileText, "utf8", error => {
 				if (error) {
 					console.error(error);
 				}
@@ -368,8 +240,8 @@ exports.clubListBuilder = function () {
 			resolve(messageOptions);
 		}).then(messageOptions => {
 			messageOptions.files = [{
-				attachment: "data/ClubChannels.txt",
-				name: "ClubChannels.txt"
+				attachment: "data/listMessage.txt",
+				name: `${listType}List.txt`
 			}];
 			messageOptions.embeds = [];
 			return messageOptions;
@@ -378,7 +250,7 @@ exports.clubListBuilder = function () {
 		return new Promise((resolve, reject) => {
 			messageOptions.embeds = [
 				embedTemplateBuilder("#f07581")
-					.setTitle("Clubs")
+					.setTitle(`${listType.toUpperCase()} LIST`)
 					.setDescription(description)
 			];
 			messageOptions.files = [];
@@ -387,129 +259,54 @@ exports.clubListBuilder = function () {
 	}
 }
 
-/** Pin the club list message
+/** Update the club or petition list message
  * @param {GuildChannelManager} channelManager
- * @param {TextChannel} channel
+ * @param {"petition" | "club"} listType
+ * @returns {Promise<Message>}
  */
-exports.pinClubList = function (channelManager, channel) {
-	exports.clubListBuilder(channelManager).then(messageOptions => {
+exports.updateList = async function (channelManager, listType) {
+	const { channelId, messageId } = exports.listMessages[listType];
+	if (channelId && messageId) {
+		const channel = await channelManager.fetch(channelId);
+		const message = await channel.messages.fetch(messageId);
+		const messageOptions = await exports.buildListMessagePayload(channelManager, listType);
+		message.edit(messageOptions);
+		if (messageOptions.files.length === 0) {
+			message.removeAttachments();
+		}
+		return message;
+	}
+}
+
+/** Pins the specified list
+ * @param {ChannelManager} channelManager
+ * @param {TextChannel} channel
+ * @param {"petition" | "club"} listType
+ */
+exports.pinList = function (channelManager, channel, listType) {
+	exports.buildListMessagePayload(channelManager, listType).then(messageOptions => {
 		channel.send(messageOptions).then(message => {
-			exports.listMessages.clubs = {
+			exports.listMessages[listType] = {
 				"messageId": message.id,
 				"channelId": message.channelId
 			}
-			message.pin();
 			exports.saveObject(exports.listMessages, "listMessageIds.json");
+			message.pin();
 		})
 	}).catch(console.error);
 }
 
-/** Create a topic channel for a petition if it has enough ids
- * @param {Guild} guild
- * @param {string} topicName
- * @param {User} author
- * @returns {{petitions: number, threshold: number}}
- */
-exports.checkPetition = function (guild, topicName, author = null) {
-	let petitions = exports.getPetitions();
-	if (!petitions[topicName]) {
-		petitions[topicName] = [];
-	}
-	if (author) {
-		if (!petitions[topicName].includes(author.id)) {
-			petitions[topicName].push(author.id);
-		} else {
-			author.send(`You have already petitioned for ${topicName}.`)
-				.catch(console.error)
-			return;
-		}
-	}
-	const petitionCount = petitions[topicName].length ?? 0;
-	const threshold = Math.ceil(guild.memberCount * 0.05) + 1;
-	if (petitionCount >= threshold) {
-		exports.addTopicChannel(guild, topicName);
-	} else {
-		exports.setPetitions(petitions, guild.channels);
-	}
-	exports.updateList(guild.channels, "topics");
-	return {
-		petitions: petitionCount,
-		threshold
-	}
-}
-
-/** Create a topic channel
- * @param {Guild} guild
- * @param {string} topicName
- * @returns {Promise<TextChannel>}
- */
-exports.addTopicChannel = function (guild, topicName) {
-	return guild.channels.create({
-		name: topicName,
-		parent: "581886288102424592",
-		permissionOverwrites: [
-			{
-				id: guild.client.user.id,
-				allow: [PermissionsBitField.Flags.ViewChannel],
-				type: 1
-			},
-			{
-				id: exports.modRoleId,
-				allow: [PermissionsBitField.Flags.ViewChannel],
-				type: 0
-			},
-			{
-				id: guild.id,
-				deny: [PermissionsBitField.Flags.ViewChannel],
-				type: 0
-			}
-		],
-		type: ChannelType.GuildText
-	}).then(channel => {
-		const petitions = exports.getPetitions();
-		if (!petitions[topicName]) {
-			petitions[topicName] = [];
-		}
-
-		// Make channel viewable by petitioners, and BountyBot
-		guild.members.fetch({
-			user: petitions[topicName].concat(["536330483852771348"])
-		}).then(allowedCollection => {
-			allowedCollection.mapValues(member => {
-				channel.permissionOverwrites.create(member.user, {
-					[PermissionsBitField.Flags.ViewChannel]: true
-				});
-			})
-
-			if (petitions[topicName].length > 0) {
-				channel.send(`This channel has been created thanks to: <@${petitions[topicName].join('> <@')}>`);
-			}
-			delete petitions[topicName];
-			addTopic(channel.id, channel.name);
-			exports.saveObject(getTopicIds(), 'topicList.json');
-			exports.setPetitions(petitions, guild.channels);
-		})
-		return channel;
-	}).catch(console.error);
-}
-
-/** Add the user to the topic/club channel (syncing internal tracking and permissions)
+/** Add the user to the club (syncing internal tracking and permissions)
  * @param {TextChannel} channel
  * @param {User} user
  */
 exports.joinChannel = function (channel, user) {
 	if (!user.bot) {
 		const { id, permissionOverwrites, guild, name: channelName } = channel;
-		let permissionOverwrite = permissionOverwrites.resolve(user.id);
-		if (!permissionOverwrite || !permissionOverwrite.deny.has(PermissionsBitField.Flags.ViewChannel, false)) {
-			if (getTopicIds().includes(id)) {
-				permissionOverwrites.create(user, {
-					[PermissionsBitField.Flags.ViewChannel]: true
-				}).then(() => {
-					channel.send(`Welcome to ${channelName}, ${user}!`);
-				}).catch(console.error);
-			} else if (Object.keys(exports.getClubDictionary()).includes(id)) {
-				let club = exports.getClubDictionary()[id];
+		const permissionOverwrite = permissionOverwrites.resolve(user.id);
+		if (!permissionOverwrite.deny.has(PermissionsBitField.Flags.ViewChannel, false)) {
+			const club = exports.getClubDictionary()[id];
+			if (!club) {
 				if (club.seats === -1 || club.isRecruiting()) {
 					if (club.hostId != user.id && !club.userIds.includes(user.id)) {
 						club.userIds.push(user.id);
@@ -522,7 +319,7 @@ exports.joinChannel = function (channel, user) {
 							channel.send(`Welcome to ${channelName}, ${user}!`);
 						})
 						exports.updateClubDetails(club, channel);
-						exports.updateList(guild.channels, "clubs");
+						exports.updateList(guild.channels, "club");
 						exports.updateClub(club);
 					} else {
 						user.send(`You are already in ${club.title}.`)
@@ -549,7 +346,7 @@ exports.updateClubDetails = (club, channel) => {
 		message.edit({ content: "You can send out invites with \`/club-invite\`. Prospective members will be shown the following embed:", embeds: [clubEmbedBuilder(club)], fetchReply: true }).then(detailSummaryMessage => {
 			detailSummaryMessage.pin();
 			club.detailSummaryId = detailSummaryMessage.id;
-			exports.updateList(channel.guild.channels, "clubs");
+			exports.updateList(channel.guild.channels, "club");
 			exports.updateClub(club);
 		});
 	}).catch(error => {
@@ -558,7 +355,7 @@ exports.updateClubDetails = (club, channel) => {
 			channel.send({ content: "You can send out invites with \`/club-invite\`. Prospective members will be shown the following embed:", embeds: [clubEmbedBuilder(club)], fetchReply: true }).then(detailSummaryMessage => {
 				detailSummaryMessage.pin();
 				club.detailSummaryId = detailSummaryMessage.id;
-				exports.updateList(channel.guild.channels, "clubs");
+				exports.updateList(channel.guild.channels, "club");
 				exports.updateClub(club);
 			});
 		} else {
@@ -572,62 +369,20 @@ exports.updateClubDetails = (club, channel) => {
  * @param {Guild} guild
  */
 exports.createClubEvent = function (club, guild) {
-	if (club.isRecruiting()) {
-		return guild.channels.fetch(club.voiceChannelId).then(voiceChannel => {
-			return guild.scheduledEvents.create({
-				name: club.title,
-				scheduledStartTime: club.timeslot.nextMeeting * 1000,
-				privacyLevel: 2,
-				entityType: GuildScheduledEventEntityType.Voice,
-				description: club.description,
-				channel: voiceChannel
-			})
-		}).then(event => {
-			club.timeslot.setEventId(event.id);
-			exports.updateList(guild.channels, "clubs");
-			exports.updateClub(club);
-		});
-	}
-}
-
-/** The number of ms until the timestamp, but not more than the max allowable setTimeout duration
- * @param {number} timestamp units: seconds
- */
-function clampedTimestampToMS(timestamp) {
-	return Math.min((timestamp * 1000) - Date.now(), MAX_SET_TIMEOUT);
-}
-
-/** Checks if the given unix timestamp can be scheduled to with 1 setTimeout
- * @param {number} timestamp units: seconds
- */
-function isTimestampWithinOneTimeout(timestamp) {
-	return (timestamp * 1000) - Date.now() <= MAX_SET_TIMEOUT;
-}
-
-/** Create a timeout to create a scheduled event for a club after the current event passes
- * @param {string} club
- * @param {string} clubVoiceId
- * @param {number | null} nextMeetingTimestamp
- * @param {Guild} guild
- */
-exports.scheduleClubEvent = function (clubId, clubVoiceId, nextMeetingTimestamp, guild) {
-	if (isTimestampWithinOneTimeout(nextMeetingTimestamp)) {
-		const timeout = setTimeout((timeoutClubId, timeoutGuild) => {
-			const club = exports.getClubDictionary()[timeoutClubId];
-			if (club?.isRecruiting()) {
-				exports.createClubEvent(club, timeoutGuild);
-			}
-		}, clampedTimestampToMS(nextMeetingTimestamp), clubId, guild);
-		exports.eventTimeouts[clubVoiceId] = timeout;
-	} else {
-		const timeout = setTimeout((timeoutClubId, timeoutGuild) => {
-			const club = exports.getClubDictionary()[timeoutClubId];
-			if (club) {
-				exports.scheduleClubEvent(club.id, club.voiceChannelId, club.timeslot.nextMeeting, timeoutGuild);
-			}
-		}, MAX_SET_TIMEOUT, clubId, guild);
-		exports.eventTimeouts[clubVoiceId] = timeout;
-	}
+	return guild.channels.fetch(club.voiceChannelId).then(voiceChannel => {
+		return guild.scheduledEvents.create({
+			name: club.title,
+			scheduledStartTime: club.timeslot.nextMeeting * 1000,
+			privacyLevel: 2,
+			entityType: GuildScheduledEventEntityType.Voice,
+			description: club.description,
+			channel: voiceChannel
+		})
+	}).then(event => {
+		club.timeslot.setEventId(event.id);
+		exports.updateList(guild.channels, "clubs");
+		exports.updateClub(club);
+	});
 }
 
 /** Delete the scheduled event associated with a club's next meeting
@@ -637,57 +392,59 @@ exports.scheduleClubEvent = function (clubId, clubVoiceId, nextMeetingTimestamp,
 exports.cancelClubEvent = function (club, eventManager) {
 	if (club.timeslot.eventId) {
 		eventManager.delete(club.timeslot.eventId).catch(console.error);
-		club.timeslot.eventId = null;
-	}
-	if (exports.eventTimeouts[club.voiceChannelId]) {
-		clearTimeout(exports.eventTimeouts[club.voiceChannelId]);
-		delete exports.eventTimeouts[club.voiceChannelId];
+		club.timeslot.setEventId(null);
 	}
 }
 
-/** Set a timeout to send a reminder message to the given club a day before its next meeting
+/** scheduled actions: send a reminder about the upcoming meeting to the club and create a scheduled event for the meeting after that
  * @param {string} clubId
  * @param {number | null} nextMeetingTimestamp
  * @param {ChannelManager} channelManager
  */
-exports.setClubReminder = async function (clubId, nextMeetingTimestamp, channelManager) {
-	if (nextMeetingTimestamp) {
-		const timeout = setTimeout(
+exports.scheduleClubReminderAndEvent = async function (clubId, nextMeetingTimestamp, channelManager) {
+	let timeout;
+	const msToTimestamp = (nextMeetingTimestamp - exports.timeConversion(1, "d", "s")) * 1000 - Date.now();
+	if (msToTimestamp <= MAX_SET_TIMEOUT) {
+		timeout = setTimeout(
 			async (clubId, channelManager) => {
 				const club = exports.getClubDictionary()[clubId];
 				if (club.timeslot.nextMeeting) {
-					if (isTimestampWithinOneTimeout(club.timeslot.nextMeeting - exports.timeConversion(1, "d", "s"))) {
-						if (club.timeslot.periodCount) {
-							await exports.sendClubReminder(club, channelManager);
-							const timeGap = exports.timeConversion(club.timeslot.periodCount, club.timeslot.periodUnits === "weeks" ? "w" : "d", "s");
-							club.timeslot.setNextMeeting(club.timeslot.nextMeeting + timeGap);
-							exports.updateClub(club);
-							exports.scheduleClubEvent(club.id, club.voiceChannelId, club.timeslot.nextMeeting, channelManager.guild);
-							exports.setClubReminder(club.id, club.timeslot.nextMeeting, channelManager);
-						} else {
-							club.timeslot.setEventId(null);
-							exports.updateList(channelManager, "clubs");
-							exports.updateClub(club, channelManager);
+					await exports.sendClubReminder(clubId, channelManager);
+					if (club.timeslot.periodCount && club.timeslot.periodUnits) {
+						const nextTimestamp = club.timeslot.nextMeeting + exports.timeConversion(club.timeslot.periodCount, club.timeslot.periodUnits === "weeks" ? "w" : "d", "s");
+						club.timeslot.setNextMeeting(nextTimestamp);
+						exports.updateClub(club);
+						if (club?.isRecruiting()) {
+							exports.createClubEvent(club, channelManager.guild);
 						}
+						exports.scheduleClubReminderAndEvent(clubId, nextTimestamp, channelManager);
 					} else {
-						exports.setClubReminder(club.id, club.timeslot.nextMeeting, channelManager);
+						delete exports.reminderTimeouts[clubId];
+						club.timeslot.setNextMeeting(null);
+						exports.cancelClubEvent(club, channelManager.guild.scheduledEvents);
+						updateClub(club);
 					}
+					exports.updateList(channelManager, "club");
 				}
 			},
-			clampedTimestampToMS(nextMeetingTimestamp - exports.timeConversion(1, "d", "s")),
+			msToTimestamp,
 			clubId,
 			channelManager);
-		exports.reminderTimeouts[clubId] = timeout;
-		exports.updateList(channelManager, "clubs");
+	} else {
+		timeout = setTimeout(() => {
+			exports.scheduleClubReminderAndEvent(clubId, nextMeetingTimestamp, channelManager);
+		}, MAX_SET_TIMEOUT);
 	}
+	exports.reminderTimeouts[clubId] = timeout;
 }
 
 /** Send a club reminder message
- * @param {Club} club
+ * @param {string} clubId
  * @param {ChannelManager} channelManager
  */
-exports.sendClubReminder = async (club, channelManager) => {
-	const textChannel = await channelManager.fetch(club.id);
+exports.sendClubReminder = async (clubId, channelManager) => {
+	const club = exports.getClubDictionary()[clubId];
+	const textChannel = await channelManager.fetch(clubId);
 	// NOTE: defaultReminder.length (without interpolated length) must be less than or equal to 55 characters so it fits in the config modal placeholder with its wrapper (100 characters)
 	const defaultReminder = `Reminder: This club will meet at <t:${club.timeslot.nextMeeting}> in <#${club.voiceChannelId}>!`;
 	const reminderPayload = {
@@ -699,7 +456,7 @@ exports.sendClubReminder = async (club, channelManager) => {
 			reminderPayload.components = [new ActionRowBuilder({
 				components: [
 					new ButtonBuilder({
-						customId: 'startevent',
+						customId: `startevent${SAFE_DELIMITER}${club.timeslot.eventId}`,
 						label: "Start Event",
 						emoji: "👑",
 						style: ButtonStyle.Primary,
@@ -715,7 +472,7 @@ exports.sendClubReminder = async (club, channelManager) => {
  * @param {string} channelId
  */
 exports.clearClubReminder = async function (channelId) {
-	if (exports.reminderTimeouts[channelId]) {
+	if (channelId in exports.reminderTimeouts) {
 		clearTimeout(exports.reminderTimeouts[channelId]);
 		delete exports.reminderTimeouts[channelId];
 	}